Implemented the "fetch assignments" method in the assignments cubit. Also reorganized the repos by user and assignment, instead of by user role.
The interfaces for the repos still need functions added, so
add what you need and make a pull request.

There are currently two cubits, one for users and one
for assignments. I am of the opinion that we should have more.
For instance, one cubit per use case. We can vote on that later.